requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
dataclasses-json==0.5.13
decorator==5.1.1
Django==4.2.3
djangorestframework==3.14.0
fastapi==0.99.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.247
langsmith==0.0.15
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
overrides==7.3.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
posthog==3.0.1
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rich==13.4.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tldextract==3.4.4
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
whitenoise==6.5.0
yarl==1.9.2
zipp==3.16.2
zope.event==4.6
zope.interface==6.0
